Streamlit_Frontend/requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.1
numpy==1.24.1
streamlit==1.16.0
streamlit-echarts==0.4.0 
requests==2.28.1 
altair==4.1.0
google.generativeai
plotly
python-dotenv 
